third_party/WebKit/LayoutTests/fast/css/invalid-font-face-crash.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style id="styleElement">
@font-face {}
</style>
<script>
test(() => {
  let fontFaceRule = styleElement.sheet.rules[0];
  assert_equals(fontFaceRule.style.getPropertyValue('font-display'), "");
}, "This test should not crash");
</script>
